Task 146 - move save SaveSearch logic to seperate class
