+++ plugins/org.eclipse.osee.orcs.rest/src/org/eclipse/osee/orcs/rest/internal/OrcsApplication.java
@@ -93,6 +93,7 @@ public class OrcsApplication extends Application {
       singletons.add(new UserPreferencesEndpointImpl(orcsApi));
       singletons.add(new UserEndpointImpl(orcsApi));
       singletons.add(new KeyValueEndpointImpl(orcsApi));
+      singletons.add(new SavedSearchEndpoint(orcsApi));
    }
 
    public void stop() {
