+++ web/apps/osee/angular.json
@@ -656,7 +656,10 @@
 			"environment": "all"
 		},
 		"packageManager": "pnpm",
-		"schematicCollections": ["@angular-eslint/schematics"]
+    "schematicCollections": [
+      "@angular-eslint/schematics"
+    ],
+    "analytics": false
 	},
 	"schematics": {
 		"@angular-eslint/schematics:application": {
